Added private method appendLayers to append weights of splitted layers.
The AlexNet weights file in HDF5 format has splitted up the weights of some layers since the original implementation of the AlexNet used two graphics cards for the calculations. The method will simply take the weights of the second half, appends it to the first one and resets the dimension to the right value.