added test data to check convolution implementation.
Data is an export from tensorflow rounded to a precision of four digits after the comma.
Tests need to implement evaluation with an epsilon corresponding to this inaccuracy.
However it should be clear very quickly, what has been calculated correctly and what not since we'll more likely fuck up index aritmetic and thus will end up with totally different results.