Added functional placement algorithm to PlatformPlacer.
According to OperationMode the PlatformPlacer chooses the platform that best fits the needs of that mode and sequentially assigns
tasks to either the perfomance platform or a fallback platform that is usually a CPU.

It does not make sense at the current point to dispatch tasks to more than 2 platforms, because processing is sequential anyways and
we already chose the 'perfect' platform for the specific operation mode.
A computation Distribution attribute has been added to the ImageResult such that the GUI can display the devices used and their percentage.

Future features could include:
- Giving layers a parameter of "difficulty", so we can decide where it makes sense to dispatch to a perfomance platform on case to case basis.
- Adding some sort of knowledge about strenghts and weaknesses of a platform to dispatch computations according to design of a platform.