fix GEMM and write tests
This commit contains the changes necessary to test an entire run of an efficient convolution (on a CPU). First, the image is reconstructed using im2col. Thanks to the reordering, a single matrix multiplication is suffice to convolve the image with kernels (that are also reordered into a weight matrix). Lastly, the computed result needs to be reordered again to the final image using col2im.

For that, I fixed the im2col algorithm, implemented an algorithm that multiplies two matrices that are represented using 1d vectors, rewrote the col2im algorithm, and wrote test cases for all that (which will be refined during the QA phase).

The main.cpp can be used for debugging and printing the output of the performed computations to the console.