Fix excludes with newer gcov versions
With gcov 1.13, the existing excludes are not working properly.
One has to use '*/tests/*' for a match to be successful or an absolute
path like '${CMAKE_SOURCE_DIR}/tests/*'. The latter is a tad more
correct, so use that.
While at it, also add '${CMAKE_BINARY_DIR}/*', so we ignore the build
directory even if it doesn't match cmake-build-*.