Added difficulty attribute to PlatformPlacer
The layers are distributed to perfomance platform or fallback platform according to their
difficulty in relation to the average, since it only makes sense to compute fairly difficulty layers
on the perfomance platform (overhead in OpenCL, missing kernels, ...).
Furthermore computation distribution is calculated based on the total difficulty on platform is 'responsible' for.