Mark Qt as REQUIRED
There are currently too many places where Qt is used:
- src: gui, manager, main.cpp
- tests: gui, manager, executor

Instead of making the use of Qt conditional, mark it as REQUIRED so we
fail early, if it isn't installed.
This means though, that compiling hics for the FPGA requires Qt even
when we don't use the GUI later on.