Add generic ClPlatform for GPU and CL_CPU
Add a separte ClPlatform (and ClConvolutionFunction) which handles GPU
and CL_CPU. As GPU and CL_CPU are similar enough, we have a single class
handling both.
Update the build system to build the FPGA code only when targetting the
FPGA board by setting the PLATFORM_ALTERA=ON config option in which case
the GPU/CL_CPU code is skipped.