Refactor method to check OpenCL errors
The OpenCL error code alone is not very helpful. It does not return a
helpful error message that could be used  for throwing exceptions. Thus
we extend the CheckError function to include a second, more descriptive
user-defined error message in addition to the original OpenCL error
message.

Furthermore, the CheckError function is designed as a template function,
so we can also pass the information which kind of exception should be
thrown if a OpenCL function was not successful.