trim aggregated results after sorting them by probability
After the results were aggregated and each label got its new
probability, the labels were trimmed to the top 5 labels in the vector
and then sorted by probability. This resulted in returning only the 5
labels of the first image because they were inserted in the first 5
fields of the vector. Of course they need to be sorted by probability
first and then trimmed to the 5 most probable results.